move safeCastChunkSize to the utils package
